# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Kata Blog API
  version: 0.2.2
  summary: REST API for the Kata blog application
  description: |
    HTTP API for the Kata blog platform. Manages users, blog posts, categories,
    tags, and images. All endpoints are mounted under the `/api` prefix.

    Conventions:
      - **Get one**: `GET /{resource}/get?by=id|slug&value=...`
      - **List**:   `GET /{resource}/query?sort=...&order=asc|desc&...`
      - **Auth**:   Bearer JWT on all write/delete endpoints.
      - **Images**: Looked up by `id` only. Upload via `multipart/form-data` to
        Cloudinary — single file at `POST /images`, up to 10 files at
        `POST /images/bulk`.

    Backed by PostgreSQL via Prisma in production; this spec is the source of
    truth for handlers and clients.
  contact:
    name: Kata
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://kata.vercel.app/api
    description: Vercel deployment (replace with your domain)
...
paths:
...
  /images/get:
    get:
      tags: [Images]
      summary: Get a single image
      operationId: getImage
      parameters:
        - $ref: "#/components/parameters/IdByParam"
        - $ref: "#/components/parameters/IdValueParam"
      responses:
        "200":
          description: The image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Image"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /images/query:
    get:
      tags: [Images]
      summary: List images
      operationId: listImages
      parameters:
        - in: query
          name: sort
          schema:
            type: string
            enum: [name, createdAt, updatedAt]
            default: createdAt
        - $ref: "#/components/parameters/OrderParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: uploaderId
          schema:
            type: string
            format: uuid
        - in: query
          name: mimeType
          schema:
            type: string
          description: Filter by exact mime type, e.g. `image/png`
      responses:
        "200":
          description: Paginated list of images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageListResponse"
  /images/bulk:
    post:
      tags: [Images]
      summary: Upload multiple images
      description: |
        Uploads up to 10 image files to Cloudinary and stores their metadata in
        the database. Requires the `author`, `editor`, or `admin` role. The
        `uploaderId` is taken from the authenticated session.

        If any file fails validation or upload, the entire request is rolled
        back and no images are saved.

        Accepted file types: JPEG, PNG, WebP, GIF. Maximum file size per file:
        10 MB.
      operationId: uploadImages
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ImageBulkUpload"
      responses:
        "201":
          description: Images uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageBulkUploadResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          description: Cloudinary upload failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /images:
    post:
      tags: [Images]
      summary: Upload a single image
      description: |
        Uploads one image file to Cloudinary and stores its metadata in the
        database. The `uploaderId` is taken from the authenticated session.

        For multiple files in one request, use `POST /images/bulk` instead.

        Accepted file types: JPEG, PNG, WebP, GIF. Maximum file size: 10 MB.
      operationId: uploadImage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ImageUpload"
      responses:
        "201":
          description: Image uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Image"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          description: Cloudinary upload failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      tags: [Images]
      summary: Update image metadata
      description: |
        Updates optional display metadata. Requires the `author`, `editor`, or
        `admin` role.
      operationId: updateImage
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdByParam"
        - $ref: "#/components/parameters/IdValueParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageUpdate"
      responses:
        "200":
          description: Updated image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Image"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Images]
      summary: Delete an image
      description: |
        Deletes the image from Cloudinary and removes its database record.
        Requires the `author`, `editor`, or `admin` role.
      operationId: deleteImage
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdByParam"
        - $ref: "#/components/parameters/IdValueParam"
      responses:
        "204":
          description: Image deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Cloudinary delete failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

# ─────────────────────────────────────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer token. The full JWT flow is wired in a later step; this spec
        declares the contract so write endpoints are protected.
  parameters:
...
    IdByParam:
      in: query
      name: by
      required: true
      schema:
        type: string
        enum: [id]
      description: Images are looked up by id only
    IdValueParam:
      in: query
      name: value
      required: true
      schema:
        type: string
        format: uuid
      description: Image UUID
...
    OrderParam:
      in: query
      name: order
      schema:
        type: string
        enum: [asc, desc]
        default: desc
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated user lacks the required role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: BLOG_NOT_FOUND }
            message: { type: string }
            details:
              type: object
              additionalProperties: true

    # ── Users ──────────────────────────────────────────────────────────────
...
    Image:
      type: object
      required: [id, url, mimeType, size, uploaderId, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        alt: { type: string }
        url:
          type: string
          format: uri
          description: Cloudinary secure URL
        cloudinaryPublicId:
          type: string
          description: Cloudinary public ID used for storage cleanup on delete
        mimeType: { type: string, example: image/png }
        size: { type: integer, minimum: 0, description: Size in bytes }
        width: { type: integer, minimum: 0 }
        height: { type: integer, minimum: 0 }
        uploaderId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ImageUpload:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: Image file (JPEG, PNG, WebP, or GIF; max 10 MB)
        name:
          type: string
          description: Optional display name (defaults to the original filename)
        alt:
          type: string
          description: Optional alt text for accessibility
    ImageBulkUpload:
      type: object
      required: [files]
      properties:
        files:
          type: array
          minItems: 1
          maxItems: 10
          items:
            type: string
            format: binary
          description: |
            One or more image files. Repeat the `files` field for each upload
            (JPEG, PNG, WebP, or GIF; max 10 MB each).
    ImageBulkUploadResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/Image" }
          description: Uploaded images, in the same order as the request files
    ImageUpdate:
      type: object
      minProperties: 1
      properties:
        name: { type: string }
        alt: { type: string }
    ImageListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Image" }
        pagination: { $ref: "#/components/schemas/Pagination" }

    # ── Blogs ──────────────────────────────────────────────────────────────
...